Keep integration tests all in test_integration itself
