+++ bowtie/tests/test_benchmarks.py
@@ -1,12 +1,9 @@
 from pathlib import Path
-from textwrap import dedent
 import json
-import uuid
 
 import pytest
 
 from bowtie import _benchmarks, benchmarks
-from bowtie._cli import EX
 from bowtie._core import Dialect, TestCase
 from bowtie._direct_connectable import Direct
 from bowtie.tests.test_integration import bowtie
@@ -218,239 +215,3 @@ class TestBenchmarker:
             valid_single_benchmark.serializable(),
             **benchmarker_run_args,
         )
-
-
-class TestBenchmarkRun:
-
-    @pytest.mark.asyncio
-    async def test_nonexistent_benchmark_run(self):
-        random_name = uuid.uuid4().hex
-        _, stderr = await bowtie(
-            "perf",
-            "-i",
-            DIRECT_CONNECTABLE,
-            "-b",
-            random_name,
-            exit_code=EX.DATAERR,
-        )
-        assert "Benchmark File not found" in stderr
-
-    @pytest.mark.asyncio
-    async def test_benchmark_run_json_output(
-        self,
-        valid_single_benchmark,
-        tmp_path,
-    ):
-        tmp_path.joinpath("benchmark.json").write_text(
-            json.dumps(valid_single_benchmark.serializable()),
-        )
-        stdout, stderr = await bowtie(
-            "perf",
-            "-i",
-            DIRECT_CONNECTABLE,
-            "-q",
-            "--format",
-            "json",
-            tmp_path / "benchmark.json",
-            exit_code=0,
-            json=True,
-        )
-        assert benchmark_report_validator.is_valid(stdout)
-
-    @pytest.mark.asyncio
-    async def test_benchmark_run_pretty_output(
-        self,
-        valid_single_benchmark,
-        tmp_path,
-    ):
-        tmp_path.joinpath("benchmark.json").write_text(
-            json.dumps(valid_single_benchmark.serializable()),
-        )
-        stdout, stderr = await bowtie(
-            "perf",
-            "-i",
-            DIRECT_CONNECTABLE,
-            "-q",
-            "--format",
-            "pretty",
-            tmp_path / "benchmark.json",
-            exit_code=0,
-        )
-
-        # FIXME: We don't assert against the exact output yet, as it's a WIP
-        assert stdout, stderr
-
-    @pytest.mark.asyncio
-    async def test_benchmark_run_markdown_output(
-        self,
-        valid_single_benchmark,
-        tmp_path,
-    ):
-        tmp_path.joinpath("benchmark.json").write_text(
-            json.dumps(valid_single_benchmark.serializable()),
-        )
-        stdout, stderr = await bowtie(
-            "perf",
-            "-i",
-            DIRECT_CONNECTABLE,
-            "-q",
-            "--format",
-            "markdown",
-            tmp_path / "benchmark.json",
-            exit_code=0,
-        )
-
-        assert (
-            dedent(
-                """\
-            # Benchmark Summary
-            ## Benchmark Group: benchmark
-            Benchmark File: None
-
-
-            Benchmark: Tests with benchmark
-
-            | Test Name | python-jsonschema |
-            """,
-            )
-            in stdout
-        )
-
-        assert (
-            dedent(
-                """\
-            ## Benchmark Metadata
-
-            Runs: 3
-            Values: 2
-            Warmups: 1
-            """,
-            )
-            in stdout
-        )
-
-    @pytest.mark.asyncio
-    async def test_benchmark_run_varying_param_markdown(
-        self,
-        tmp_path,
-    ):
-        from bowtie.tests.benchmarks import benchmark_with_varying_parameter
-
-        tmp_path.joinpath("benchmark.json").write_text(
-            json.dumps(
-                benchmark_with_varying_parameter.get_benchmark().serializable(),
-            ),
-        )
-        stdout, stderr = await bowtie(
-            "perf",
-            "-i",
-            DIRECT_CONNECTABLE,
-            "-q",
-            "--format",
-            "markdown",
-            tmp_path / "benchmark.json",
-            exit_code=0,
-        )
-
-        assert (
-            dedent(
-                """\
-            # Benchmark Summary
-            ## Benchmark Group: benchmark
-            Benchmark File: None
-
-            Benchmark: Tests with varying Array Size
-
-            | Test Name | python-jsonschema |
-            """,
-            ).strip("\n")
-            in stdout
-        )
-
-        assert (
-            dedent(
-                """\
-            Benchmark: Tests with benchmark 2
-
-            | Test Name | python-jsonschema |
-            """,
-            )
-            in stdout
-        )
-
-        assert (
-            dedent(
-                """\
-            ## Benchmark Metadata
-
-            Runs: 3
-            Values: 2
-            Warmups: 1
-            """,
-            )
-            in stdout
-        )
-
-    @pytest.mark.asyncio
-    async def test_invalid_benchmark_run(self, invalid_benchmark, tmp_path):
-        tmp_path.joinpath("benchmark.json").write_text(
-            json.dumps(invalid_benchmark),
-        )
-        _, stderr = await bowtie(
-            "perf",
-            "-i",
-            DIRECT_CONNECTABLE,
-            "-q",
-            "--format",
-            "json",
-            tmp_path / "benchmark.json",
-            exit_code=1,
-        )
-
-
-class TestFilterBenchmarks:
-
-    @pytest.mark.parametrize(
-        "dialect",
-        Dialect.known(),
-        ids=lambda param: param.short_name,
-    )
-    @pytest.mark.asyncio
-    async def test_default_benchmarks(self, dialect):
-        stdout, stderr = await bowtie(
-            "filter-benchmarks",
-            "-D",
-            dialect.short_name,
-            exit_code=0,
-        )
-        assert stderr == ""
-
-    @pytest.mark.parametrize(
-        "dialect",
-        Dialect.known(),
-        ids=lambda param: param.short_name,
-    )
-    @pytest.mark.asyncio
-    async def test_keyword_benchmarks(self, dialect):
-        stdout, stderr = await bowtie(
-            "filter-benchmarks",
-            "-D",
-            dialect.short_name,
-            "-t",
-            "keyword",
-            exit_code=0,
-        )
-        assert stderr == ""
-
-    @pytest.mark.asyncio
-    async def test_filtering_by_name(self):
-        stdout, stderr = await bowtie(
-            "filter-benchmarks",
-            "-t",
-            "keyword",
-            "-n",
-            uuid.uuid4().hex,
-            exit_code=0,
-        )
-        assert stdout == ""
-        assert stderr == ""
